requirements.txt
farm-haystack
streamlit
torch
pytube
llama-cpp-python
cached_property